requirements.txt
dnspython
requests
beautifulsoup4
python-whois